requirements.txt
fastapi==0.111.0
uagents==0.12.2
google-generativeai==0.5.4
pathlib==1.0.1
